#156: use deleteReply to consolidate readymove flow
